Fix: Add 'run client ()' prefix for network queries
When headers specify 'destination: network', wrap the SQL command with 'run client ()'
to route the query to operator nodes in the EdgeLake network instead of trying to
execute locally on the query node.

This fixes the error: 'Logical DBMS not opened on Operator Node' by ensuring
distributed queries are sent to the correct nodes that host the data.